Multiplicity rolled back from the profiles.
Also the number of detector resolution configurations
has been increased to incorporate T0 resolution obtained
from V0 and TPC or SPD.

The way configurations is handled is modified to make it
more rational. Also the name of the correlation profiles
are made unique now that we have crossed detector presence.